+++ app/src/main/java/com/nickoehler/brawlhalla/clans/domain/ClanMember.kt
@@ -3,9 +3,9 @@ package com.nickoehler.brawlhalla.clans.domain
 import java.time.LocalDateTime
 
 data class ClanMember(
-    val brawlhallaId: Int,
+    val brawlhallaId: Long,
     val name: String,
     val rank: ClanRankType,
     val joinDate: LocalDateTime,
-    val xp: Int
+    val xp: Long
 )
