+++ app/src/main/java/com/nickoehler/brawlhalla/legends/domain/LegendsDataSource.kt
@@ -5,5 +5,5 @@ import com.nickoehler.brawlhalla.core.domain.util.Result
 
 interface LegendsDataSource {
     suspend fun getLegends(): Result<List<Legend>, NetworkError>
-    suspend fun getLegendDetail(legendId: Int): Result<LegendDetail, NetworkError>
+    suspend fun getLegendDetail(legendId: Long): Result<LegendDetail, NetworkError>
 }
