+++ app/src/main/java/com/nickoehler/brawlhalla/ranking/presentation/StatDetailAction.kt
@@ -7,11 +7,11 @@ import com.nickoehler.brawlhalla.ranking.presentation.models.StatFilterType
 import com.nickoehler.brawlhalla.ranking.presentation.models.StatType
 
 sealed interface StatDetailAction {
-    data class SelectPlayer(val brawlhallaId: Int) : StatDetailAction
+    data class SelectPlayer(val brawlhallaId: Long) : StatDetailAction
     data class SelectStatType(val stat: StatType) : StatDetailAction
     data class SelectRankingFilterType(val type: RankingFilterType) : StatDetailAction
     data class SelectStatFilterType(val type: StatFilterType) : StatDetailAction
-    data class TogglePlayerFavorites(val brawlhallaId: Int, val name: String) : StatDetailAction
+    data class TogglePlayerFavorites(val brawlhallaId: Long, val name: String) : StatDetailAction
     data class SelectRankingModalType(val modalType: RankingModalType?) : StatDetailAction
     data class SortBy(val sortType: RankingSortType) : StatDetailAction
     data object StatLegendSortTypeReversed : StatDetailAction
